Move Scene.update_skipping_status()
